Rename pattern from `serverSetting.` to `querySetting` for consistency
Server Settings are easy to be confused with global system-wide settings while users only have the mean to rename connection settings.
In V3 Connect API the concept of query level settings is called "querySettings", thus we propose the same name in JDBC.
We still accept the old name for transition purposes but will log a warning message.

Drive-By:
 - Factor our the test Hyper Server Process management into separate class for clarity
 - Improve shutdown speed by using waitFor instead of fixed 3s sleep
 - Be strict on config settings as previously we had a silently a setting that hyper wouldn't recognize
 - Use auto port picking for the server process to avoid accidental colissions
 - Be consciously less robust on server process management during test to make any failures and problems very explicitly