Adjust `fromChannel` method to minimize injected headers and add examples
The intent of this change was to analyze how to use the JDBC driver in a least invasive way (compared to manual gRPC calling).
To reach that minimum the `DataCloudConnection.fromChannel` method was adjusted to only add the must have interceptors, which - without usage of the corresponding connection parameters - are complete no ops.

To demonstrate this flow a new `examples` test package was create that demonstrates the approach and serves as documentation.
To write that test case (and minimize the use of helpers which would hide complexity) the test hyper instance variable with its getPort method was exposed as public.

Drive-By:t
- The hyper test instance was made non-static to avoid problems in parallel test execution.
- A sample implementation of using the row based API to serve row based paginated results was also added and commented. With the parallel ongoing work we should see simplification of it.