Remove `listener` from `StreamingResultSet`
This commit de-obfuscates the code inside the `StreamingResultSet`.

The `listener` was always set to an `AlreadyReadyNoopListener` which is
essentially nothing else than a wrapper around the query-id. All its
other variables (e.g., `ready`) were hardcoded. Maybe this abstraction
served some purpose at some earlier time, but by now it was only
obfuscating the code.

This commit removes that code obfuscation.

After deobfuscation, it became obvious that `DataCloudResultSet.isReady`
was essentially hard-coded to true. As such, I am also removing that
function in this commit.