fix: Switch build back to Gradle 8
This reverts partially commit 541cf72cf30776279d62a19b2f410d5c730ed6ef.
Gradle 9 is incompatible with VS Code IDE integration and it's unclear whether it is due to a bug in Gradle or on the IDE side.
https://github.com/microsoft/vscode-gradle/issues/1712. We should wait with Gradle 9 adoption until it stabilize.

The reverted commit also did drive-by changes, this change keeps those:
- The `.gitattributes` is kept
- The updated modulegraph and freefair lombok dependencies are kept
- For gradle up shadow it was using a very early 9.0.2 version and this change uses an more recent version as there were several critical fixes since then