chore: Automate 3PP version updates
We had quite a backlog of outdated versions. This change adds the `version-catalog-update` Gradle plugin to automate the update process.
As we recently had problems with reproducible builds due to a 3PP minor update having broken our build this change also switches all libraries to exact versioning.
For the special cases where we can't update due to Java 8 / Spark or just migration concerns the logic was explicitly added to block such version candidates.

Drive-By:
- The `DEVELOPMENT.md` is updated to document this (and the existing module graph) commands that should be regularly run to have a healthy code base.
- Explicitly specify Netty version to avoid insecure version pulled in through transitive dependencies
- The updated version of Guava removed the `findbugs:jsr305` transitive dependency and thus the `@Immutable` doc annotation got broken in `EncodingUtils`. As we don't use the annotation anywhere else, we removed it